.setup/requirements.txt
# Core dependencies
flask>=2.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
pyyaml>=6.0.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-html>=3.0.0
coverage>=6.0.0

# Documentation dependencies
mkdocs>=1.2.0
mkdocs-material>=8.0.0
plantuml>=0.3.0

# AWS dependencies
boto3>=1.20.0
awscli>=1.22.0

# Development dependencies
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
isort>=5.10.0
pylint>=2.12.0
